DeVerify-Scrapping/requirements.txt
playwright==1.40.0
pymongo==4.4.0
python-dotenv==1.0.0
